pyproject.toml
[project]
name = "langgraphbasics"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "langchain-tavily>=0.2.1",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "unstructured[all-docs]>=0.17.2",
]
